Removed methods that are unused and not needed
Removed the haspaid?, correct_case?, ign_is_not_skull and
ign_is_not_mojang methods because they aren't used anymore, since we now
have a connection between the server and the website to check if they
have paid and that the skull is a skull. I also removed the
account_exists? method because it makes problems when you setup the
database since some users can be seen to not exist.