An Option to resend the confirmation mail
You should now be able to resend the confirmation mail with the click of
a link that is next to the warning that the mail isn't confirmed.

Resend the confirmation mail cleaning

Just a way to re use the files and stuff we already have instead of the
file I created even though I could use the register_mail we have.