An Option to resend the confirmation mail
You should now be able to resend the confirmation mail with the click of
a link that is next to the warning that the mail isn't confirmed.

Resend the confirmation mail cleaning

Just a way to re use the files and stuff we already have instead of the
file I created even though I could use the register_mail we have.

Change of route and link to button

I changed the route from GET to POST because of security reasons, and
changed the link_to to a button_to

changed the notice

I changed the notice so it said check for the mail instead of check for
the link

Changed notice and button

Changed the notice to say "Check your inbox" instead of "Check your
mail" also changed the way the class looks